chore: simplify test support code
