+++ datasketches/tests/countmin_serialization_test.rs
@@ -17,14 +17,15 @@
 
 #![cfg(feature = "countmin")]
 
-mod common;
+#[path = "support/serialization_data.rs"]
+mod serialization_data;
 
 use std::fs;
 
-use common::serialization_test_data;
 use datasketches::countmin::CountMinSketch;
 use googletest::assert_that;
 use googletest::prelude::contains_substring;
+use serialization_data::serialization_test_data;
 
 // This test validates binary format compatibility (deserialize + byte round-trip) for
 // C++ Count-Min snapshots. It intentionally does not assert estimate equivalence against
