+++ datasketches/tests/tuple_intersection_test.rs
@@ -17,15 +17,18 @@
 
 #![cfg(feature = "tuple")]
 
-mod common;
+#[path = "support/tuple_builder.rs"]
+mod tuple_builder;
+#[path = "support/tuple_fixtures.rs"]
+mod tuple_fixtures;
 
 use datasketches::tuple::CompactTupleSketch;
 use datasketches::tuple::SummaryCombinePolicy;
 use datasketches::tuple::SummaryPolicy;
 use datasketches::tuple::TupleIntersection;
 
-use crate::common::default_tuple_sketch_builder;
-use crate::common::tuple_sketch_with_range;
+use crate::tuple_builder::default_tuple_sketch_builder;
+use crate::tuple_fixtures::tuple_sketch_with_range;
 
 #[derive(Debug, Default, Clone, Copy)]
 struct SumPolicy;
